+++ tksheet/column_headers.py
@@ -197,8 +197,7 @@ class ColumnHeaders(tk.Canvas):
                         self.toggle_select_col(c, redraw=True)
                     if self.MT.rc_popup_menus_enabled:
                         popup_menu = self.ch_rc_popup_menu
-        if self.extra_rc_func is not None:
-            self.extra_rc_func(event)
+        try_binding(self.extra_rc_func, event)
         if popup_menu is not None:
             self.popup_menu_loc = c
             popup_menu.tk_popup(event.x_root, event.y_root)
@@ -218,10 +217,9 @@ class ColumnHeaders(tk.Canvas):
                     self.being_drawn_item = True
                     self.being_drawn_item = self.add_selection(c, set_as_current=True, run_binding_func=False)
                     self.MT.main_table_redraw_grid_and_text(redraw_header=True, redraw_row_index=True)
-                    if self.ctrl_selection_binding_func is not None:
-                        self.ctrl_selection_binding_func(
-                            self.MT.get_select_event(being_drawn_item=self.being_drawn_item),
-                        )
+                    sel_event = self.MT.get_select_event(being_drawn_item=self.being_drawn_item)
+                    try_binding(self.ctrl_selection_binding_func, sel_event)
+                    self.PAR.emit_event("<<SheetSelect>>", data=sel_event)
                 elif c_selected:
                     self.MT.deselect(c=c)
         elif not self.MT.ctrl_select_enabled:
@@ -252,10 +250,9 @@ class ColumnHeaders(tk.Canvas):
                             set_as_current=True,
                         )
                     self.MT.main_table_redraw_grid_and_text(redraw_header=True, redraw_row_index=True)
-                    if self.ctrl_selection_binding_func is not None:
-                        self.ctrl_selection_binding_func(
-                            self.MT.get_select_event(being_drawn_item=self.being_drawn_item),
-                        )
+                    sel_event = self.MT.get_select_event(being_drawn_item=self.being_drawn_item)
+                    try_binding(self.ctrl_selection_binding_func, sel_event)
+                    self.PAR.emit_event("<<SheetSelect>>", data=sel_event)
                 elif c_selected:
                     self.dragged_col = DraggedRowColumn(
                         dragged=c,
@@ -275,15 +272,14 @@ class ColumnHeaders(tk.Canvas):
                     if self.MT.selected and self.MT.selected.type_ == "columns":
                         r_to_sel, c_to_sel = self.MT.selected.row, self.MT.selected.column
                         self.MT.deselect("all", redraw=False)
-                        self.being_drawn_item = self.MT.create_selection_box(*self.get_shift_select_box(c, c_to_sel))
+                        self.being_drawn_item = self.MT.create_selection_box(*self.get_shift_select_box(c, c_to_sel), "columns")
                         self.MT.set_currently_selected(r_to_sel, c_to_sel, self.being_drawn_item)
                     else:
                         self.being_drawn_item = self.select_col(c, run_binding_func=False)
                     self.MT.main_table_redraw_grid_and_text(redraw_header=True, redraw_row_index=True)
-                    if self.shift_selection_binding_func is not None:
-                        self.shift_selection_binding_func(
-                            self.MT.get_select_event(being_drawn_item=self.being_drawn_item),
-                        )
+                    sel_event = self.MT.get_select_event(being_drawn_item=self.being_drawn_item)
+                    try_binding(self.shift_selection_binding_func, sel_event)
+                    self.PAR.emit_event("<<SheetSelect>>", data=sel_event)
                 elif c_selected:
                     self.dragged_col = DraggedRowColumn(
                         dragged=c,
@@ -371,8 +367,7 @@ class ColumnHeaders(tk.Canvas):
                         self.MT.current_cursor = "hand2"
             if not mouse_over_resize and not mouse_over_selected:
                 self.MT.reset_mouse_motion_creations()
-        if self.extra_motion_func is not None:
-            self.extra_motion_func(event)
+        try_binding(self.extra_motion_func, event)
 
     def double_b1(self, event: object):
         self.mouseclick_outside_editor_or_dropdown_all_canvases(inside=True)
@@ -412,8 +407,7 @@ class ColumnHeaders(tk.Canvas):
                     self.open_cell(event)
         self.rsz_w = None
         self.mouse_motion(event)
-        if self.extra_double_b1_func is not None:
-            self.extra_double_b1_func(event)
+        try_binding(self.extra_double_b1_func, event)
 
     def b1_press(self, event: object):
         self.MT.unbind("<MouseWheel>")
@@ -478,8 +472,7 @@ class ColumnHeaders(tk.Canvas):
                         self.being_drawn_item = self.select_col(c, redraw=True)
                     elif self.MT.toggle_selection_enabled:
                         self.toggle_select_col(c, redraw=True)
-        if self.extra_b1_press_func is not None:
-            self.extra_b1_press_func(event)
+        try_binding(self.extra_b1_press_func, event)
 
     def b1_motion(self, event: object):
         x1, y1, x2, y2 = self.MT.get_canvas_visible_area()
@@ -570,16 +563,14 @@ class ColumnHeaders(tk.Canvas):
                         else:
                             self.being_drawn_item = self.select_col(self.MT.selected.column, run_binding_func=False)
                         need_redraw = True
-                        if self.drag_selection_binding_func is not None:
-                            self.drag_selection_binding_func(
-                                self.MT.get_select_event(being_drawn_item=self.being_drawn_item),
-                            )
+                        sel_event = self.MT.get_select_event(being_drawn_item=self.being_drawn_item)
+                        try_binding(self.drag_selection_binding_func, sel_event)
+                        self.PAR.emit_event("<<SheetSelect>>", data=sel_event)
                 if self.scroll_if_event_offscreen(event):
                     need_redraw = True
             if need_redraw:
                 self.MT.main_table_redraw_grid_and_text(redraw_header=True, redraw_row_index=False)
-        if self.extra_b1_motion_func is not None:
-            self.extra_b1_motion_func(event)
+        try_binding(self.extra_b1_motion_func, event)
 
     def get_b1_motion_box(self, start_col, end_col):
         if end_col >= start_col:
@@ -631,10 +622,9 @@ class ColumnHeaders(tk.Canvas):
                             self.MT.hide_selection_box(self.MT.selected.fill_iid)
                             self.being_drawn_item = self.add_selection(box[1], run_binding_func=False)
                         need_redraw = True
-                        if self.drag_selection_binding_func is not None:
-                            self.drag_selection_binding_func(
-                                self.MT.get_select_event(being_drawn_item=self.being_drawn_item),
-                            )
+                        sel_event = self.MT.get_select_event(being_drawn_item=self.being_drawn_item)
+                        try_binding(self.drag_selection_binding_func, sel_event)
+                        self.PAR.emit_event("<<SheetSelect>>", data=sel_event)
                 if self.scroll_if_event_offscreen(event):
                     need_redraw = True
             if need_redraw:
@@ -770,10 +760,9 @@ class ColumnHeaders(tk.Canvas):
                 c_to_sel,
                 item=self.MT.create_selection_box(*to_sel, set_current=False),
             )
-            if self.drag_selection_binding_func is not None:
-                self.drag_selection_binding_func(
-                    self.MT.get_select_event(being_drawn_item=self.being_drawn_item),
-                )
+            sel_event = self.MT.get_select_event(being_drawn_item=self.being_drawn_item)
+            try_binding(self.drag_selection_binding_func, sel_event)
+            self.PAR.emit_event("<<SheetSelect>>", data=sel_event)
         self.MT.bind("<MouseWheel>", self.MT.mousewheel)
         if self.width_resizing_enabled and self.rsz_w is not None and self.currently_resizing_width:
             self.currently_resizing_width = False
@@ -879,8 +868,7 @@ class ColumnHeaders(tk.Canvas):
         self.rsz_w = None
         self.rsz_h = None
         self.mouse_motion(event)
-        if self.extra_b1_release_func is not None:
-            self.extra_b1_release_func(event)
+        try_binding(self.extra_b1_release_func, event)
 
     def toggle_select_col(
         self,
@@ -1646,7 +1634,7 @@ class ColumnHeaders(tk.Canvas):
             text = event.char
         else:
             return False
-        if self.extra_begin_edit_cell_func is not None:
+        if self.extra_begin_edit_cell_func:
             try:
                 text = self.extra_begin_edit_cell_func(
                     event_dict(